Allow more parameters to be sent to Predis Client (#6)
* Allow more parameters to be sent to Predis Client

I had a case where I needed to provide a password to connect to Redis and the current code did not allow it.
Since everything from the Suite configuration goes to the this->config array, we can simply provide it and allow the user to set the password property
This will also allow for other specific configurations

* Add description for Predis functionality

I added the full config to Predis' client, so the developer can pass username/password and/or other options. This commit adds some description above.